// frontend/src/types/Patient/Patient.ts
export type Patients = {
  firstName: string;
  middleName?: string;
  lastName: string;
  dateOfBirth: string;
  gender: string;
  phone: string;
  email: string;
  address: string;
  preferredLanguage: string;
  nationality: string;
  religion?: string;
  emergencyContact_name?: string;
  emergencyContact_relationship?: string;
};